refactor: fetch 유틸 함수 -> fetch 커스텀 훅으로 변경 및 mock url 상수화
