+++ src/components/ProductList.tsx
@@ -1,8 +1,7 @@
-import { useEffect } from 'react';
 import { atom, useRecoilState } from 'recoil';
 import { styled } from 'styled-components';
 import ProductItem from './ProductItem';
-import { fetchData } from '../utils/fetchData';
+import { useFetchData } from '../hooks/useFetchData';
 import { Product } from '../types';
 import { MOCK_DATA_URL } from '../constants/url';
 
@@ -12,14 +11,9 @@ export const productListState = atom<Product[]>({
 });
 
 const ProductList = () => {
-  const [productList, setProductList] =
-    useRecoilState<Product[]>(productListState);
+  const [productList, setProductList] = useRecoilState<Product[]>(productListState);
 
-  const productData = process.env.PUBLIC_URL + MOCK_DATA_URL;
-
-  useEffect(() => {
-    fetchData<Product[]>(productData, setProductList);
-  }, [productData, setProductList]);
+  useFetchData<Product[]>(MOCK_DATA_URL, setProductList);
 
   return (
     <S.Wrapper>
