+++ src/constants/url.ts
@@ -1,1 +1,1 @@
-export const MOCK_DATA_URL = '/mock/mockData.json'
+export const MOCK_DATA_URL = process.env.PUBLIC_URL + '/mock/mockData.json';
